Move chart-registry global functions to compat layer.
